chore: refactor reusing page logic
